follow right design pattern of service center
